+++ sdk/cpp/tests/test_sanity_codec.cpp
@@ -1,6 +1,6 @@
 /*  ----------------------------------------------------------------
  YDK - YANG Development Kit
- Copyright 2016 Cisco Systems. All rights reserved.
+ Copyright 2016-2019 Cisco Systems. All rights reserved.
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -598,4 +598,60 @@ TEST_CASE( "test_codec_action_node" )
     ydk::path::Codec s{};
     auto xml_reply = s.encode(*reply_dn, EncodingFormat::XML, false);
     REQUIRE(xml_reply==R"(<data xmlns="http://cisco.com/ns/yang/ydktest-action"><action-node><t>ok</t></action-node></data>)");
+}
+
+TEST_CASE("test_codec_augment_subtree_xml")
+{
+    CodecServiceProvider codec_provider{EncodingFormat::XML};
+    CodecService codec_service{};
+
+    auto passive = make_unique<ydktest_sanity::Runner::Passive>();
+    passive->name = "xyz";
+
+    auto ifc = make_shared<ydktest_sanity::Runner::Passive::Interfac>();
+    ifc->test = "abc";
+    passive->interfac.append(ifc);
+
+    passive->testc->xyz = make_shared<ydktest_sanity::Runner::Passive::Testc::Xyz>();
+    passive->testc->xyz->parent = passive.get();
+    passive->testc->xyz->xyz = 25;
+
+    auto xml = codec_service.encode(codec_provider, *passive, true);
+    string expected = R"(<passive xmlns="http://cisco.com/ns/yang/ydktest-sanity">
+  <name>xyz</name>
+  <interfac>
+    <test>abc</test>
+  </interfac>
+  <testc xmlns="http://cisco.com/ns/yang/ydktest-sanity-augm">
+    <xyz>
+      <xyz>25</xyz>
+    </xyz>
+  </testc>
+</passive>)";
+    CHECK(expected == xml);
+    auto entity = codec_service.decode(codec_provider, xml, make_shared<ydktest_sanity::Runner::Passive>(), true);
+    CHECK(*passive == *entity);
+}
+
+TEST_CASE("test_codec_augment_subtree_json")
+{
+    CodecServiceProvider codec_provider{EncodingFormat::JSON};
+    CodecService codec_service{};
+
+    auto passive = make_unique<ydktest_sanity::Runner::Passive>();
+    passive->name = "xyz";
+
+    auto ifc = make_shared<ydktest_sanity::Runner::Passive::Interfac>();
+    ifc->test = "abc";
+    passive->interfac.append(ifc);
+
+    passive->testc->xyz = make_shared<ydktest_sanity::Runner::Passive::Testc::Xyz>();
+    passive->testc->xyz->parent = passive.get();
+    passive->testc->xyz->xyz = 25;
+
+    auto json = codec_service.encode(codec_provider, *passive, false, true);
+    CHECK(R"({"ydktest-sanity:passive":{"interfac":[{"test":"abc"}],"name":"xyz","ydktest-sanity-augm:testc":{"xyz":{"xyz":25}}}})"
+          == json);
+    auto entity = codec_service.decode(codec_provider, json, make_shared<ydktest_sanity::Runner::Passive>(), true);
+    CHECK(*passive == *entity);
 }