+++ sdk/cpp/CHANGES.md
@@ -1,10 +1,11 @@
-### 2021-04-30 version 0.8.5.2
+### 2021-05-30 version 0.8.5.2
 
 #### Resolved GitHub issues
   * Version numbers in bundle resolver to comply with PEP 440 ([#1007](https://github.com/CiscoDevNet/ydk-gen/issues/1007))
   * The Executor Service fails to return module content to user ([#1035](https://github.com/CiscoDevNet/ydk-gen/issues/1035))
   * Docker base image python error ([#1036](https://github.com/CiscoDevNet/ydk-gen/issues/1036))
   * XmlSubtreeCodec fails encode leaf-list data ([#1045](https://github.com/CiscoDevNet/ydk-gen/issues/1045))
+  * C++ CodecService fails to encode non-top level entity to XML subtree ([#1049](https://github.com/CiscoDevNet/ydk-gen/issues/1049))
 
 #### Other resolved issues and enhancements
   * Fixed few minor bugs in C++ code